+++ src/views/app/folder-panel/messages/message-list-item-action-wrapper.tsx
@@ -6,8 +6,10 @@
 import React, { ReactNode, useMemo } from 'react';
 
 import { ContainerProps } from '@zextras/carbonio-design-system';
+import { useTranslation } from 'react-i18next';
 
 import { normalizeDropdownActionItem } from '../../../../helpers/actions';
+import { isDraft } from '../../../../helpers/folders';
 import { useMsgActions } from '../../../../hooks/actions/use-msg-actions';
 import { useTagDropdownItem } from '../../../../hooks/use-tag-dropdown-item';
 import { MailMessage } from '../../../../types';
@@ -40,6 +42,7 @@ export const MessageListItemActionWrapper = ({
 		replyDescriptor,
 		replyAllDescriptor,
 		forwardDescriptor,
+		forwardAsAttachmentDescriptor,
 		moveToTrashDescriptor,
 		deletePermanentlyDescriptor,
 		messageReadDescriptor,
@@ -63,13 +66,23 @@ export const MessageListItemActionWrapper = ({
 	} = useMsgActions({ message: item, deselectAll, shouldReplaceHistory, messagePreviewFactory });
 
 	const tagItem = useTagDropdownItem(applyTagDescriptor, item.tags);
+	const draftItem = isDraft(item.parent);
+	const [t] = useTranslation();
 
 	const dropdownItems = useMemo(
 		() =>
 			[
 				normalizeDropdownActionItem(replyDescriptor),
 				normalizeDropdownActionItem(replyAllDescriptor),
-				normalizeDropdownActionItem(forwardDescriptor),
+				{
+					id: 'ForwardMenu',
+					icon: 'Forward',
+					label: t('action.forward', 'Forward'),
+					items: [
+						normalizeDropdownActionItem(forwardDescriptor),
+						normalizeDropdownActionItem(forwardAsAttachmentDescriptor)
+					]
+				},
 				normalizeDropdownActionItem(sendDraftDescriptor),
 				normalizeDropdownActionItem(moveToTrashDescriptor),
 				normalizeDropdownActionItem(deletePermanentlyDescriptor),
@@ -90,7 +103,7 @@ export const MessageListItemActionWrapper = ({
 				normalizeDropdownActionItem(editAsNewDescriptor),
 				normalizeDropdownActionItem(showOriginalDescriptor),
 				normalizeDropdownActionItem(downloadEmlDescriptor)
-			].filter((action) => !action.disabled),
+			].filter((action) => !action.disabled && !(draftItem && action.id === 'ForwardMenu')),
 		[
 			createAppointmentDescriptor,
 			deletePermanentlyDescriptor,
@@ -99,6 +112,7 @@ export const MessageListItemActionWrapper = ({
 			editDraftDescriptor,
 			flagDescriptor,
 			forwardDescriptor,
+			forwardAsAttachmentDescriptor,
 			markAsNotSpamDescriptor,
 			markAsSpamDescriptor,
 			messageReadDescriptor,
@@ -114,7 +128,9 @@ export const MessageListItemActionWrapper = ({
 			sendDraftDescriptor,
 			showOriginalDescriptor,
 			tagItem,
-			unflagDescriptor
+			unflagDescriptor,
+			draftItem,
+			t
 		]
 	);