+++ src/views/sidebar/commons/sync-data-handler-hooks.ts
@@ -3,14 +3,15 @@
  *
  * SPDX-License-Identifier: AGPL-3.0-only
  */
-import { useEffect, useState } from 'react';
+import { useEffect, useRef, useState } from 'react';
 
-import { getTags, SoapNotify, useNotify, useRefresh } from '@zextras/carbonio-shell-ui';
+import { SoapNotify, useNotify, useRefresh } from '@zextras/carbonio-shell-ui';
 import { filter, find, forEach, isEmpty, keyBy, map, reduce, sortBy } from 'lodash';
 import { StoreApi, UseBoundStore } from 'zustand';
 
 import { useFolderStore } from '../../../carbonio-ui-commons/store/zustand/folder';
-import { folderWorker } from '../../../carbonio-ui-commons/worker';
+import { getTags, useTagStore } from '../../../carbonio-ui-commons/store/zustand/tags';
+import { folderWorker, tagsWorker } from '../../../carbonio-ui-commons/worker';
 import { useAppDispatch, useAppSelector } from '../../../hooks/redux';
 import {
 	normalizeConversation,
@@ -31,7 +32,6 @@ import {
 import {
 	handleCreatedMessages,
 	handleDeletedMessages,
-	handleModifiedMessages,
 	selectMessages
 } from '../../../store/messages-slice';
 import {
@@ -72,6 +72,7 @@ export const useSyncDataHandler = (): void => {
 	const [initialized, setInitialized] = useState(false);
 	const currentFolder = useAppSelector(selectCurrentFolder);
 	const messagesState = useAppSelector(selectMessages);
+	const processedNotify = useRef(-1);
 
 	const refresh = useRefresh();
 	useEffect(() => {
@@ -92,96 +93,99 @@ export const useSyncDataHandler = (): void => {
 		if (initialized) {
 			if (notifyList.length > 0) {
 				forEach(sortBy(notifyList, 'seq'), (notify: any) => {
-					if (!isEmpty(notify) && (notify.seq > seq || (seq > 1 && notify.seq === 1))) {
-						handleFoldersNotify(notifyList, notify, folderWorker, useFolderStore);
+					if (processedNotify.current < notify.seq) {
+						processedNotify.current = notify.seq;
+						if (!isEmpty(notify) && (notify.seq > seq || (seq > 1 && notify.seq === 1))) {
+							handleFoldersNotify(notifyList, notify, folderWorker, useFolderStore);
 
-						const tags = getTags();
-						if (notify.created) {
-							if (notify.created.c && notify.created.m) {
-								const conversations = map(notify.created.c, (i) =>
-									normalizeConversation({ c: i, m: notify.created.m, tags })
-								);
-								// eslint-disable-next-line @typescript-eslint/ban-ts-comment
-								// @ts-ignore
-								dispatch(handleNotifyCreatedConversations(conversations));
-							}
-							if (notify.created.m) {
-								dispatch(handleCreatedMessages({ m: notify.created.m }));
-								dispatch(handleCreatedMessagesInConversation({ m: notify.created.m }));
-							}
-						}
-						if (notify.modified) {
-							const soapModifiedConversations = notify.modified.c;
-							if (soapModifiedConversations) {
-								const conversations = map(soapModifiedConversations, (i) =>
-									normalizeConversation({ c: i, tags })
-								);
-								// eslint-disable-next-line @typescript-eslint/ban-ts-comment
-								// @ts-ignore
-								dispatch(handleNotifyModifiedConversations(keyBy(conversations, 'id')));
-								updateConversationsOnly(normalizeConversations(soapModifiedConversations, tags));
-							}
-							const soapModifiedMessages = notify.modified.m;
-							if (soapModifiedMessages) {
-								const messages = map(soapModifiedMessages, (obj) =>
-									normalizeMailMessageFromSoap(obj)
-								);
-								// eslint-disable-next-line @typescript-eslint/ban-ts-comment
-								// @ts-ignore
-								dispatch(handleModifiedMessages(messages));
-								updateMessagesOnly(messages);
-
-								// the condition filters messages with parent property (the only ones we need to update)
-								const toUpdate = filter(messages, 'parent');
-								if (toUpdate?.length > 0) {
-									// this function updates messages' parent in conversations. If parent never changes it does not need to be called
-									// eslint-disable-next-line @typescript-eslint/ban-ts-comment
-									// @ts-ignore
-									dispatch(handleModifiedMessagesInConversation(toUpdate));
-								}
-								// the condition filters messages with conversation property (the only ones we need to add to conversation)
-								const conversationToUpdate = filter(messages, 'conversation');
-								if (conversationToUpdate?.length > 0) {
-									const msgsReference = reduce(
-										conversationToUpdate,
-										(
-											acc: Array<{
-												id: string;
-												parent: string | undefined;
-												date: number | undefined;
-												conversation: Conversation;
-											}>,
-											msg: any
-										) => {
-											if (messagesState?.[msg?.id]) {
-												return [
-													...acc,
-													{
-														id: messagesState?.[msg?.id].id,
-														parent: messagesState?.[msg?.id].parent,
-														date: messagesState?.[msg?.id].date,
-														conversation: msg.conversation
-													}
-												];
-											}
-											return acc;
-										},
-										[]
+							const tags = getTags();
+							if (notify.created) {
+								if (notify.created.c && notify.created.m) {
+									const conversations = map(notify.created.c, (i) =>
+										normalizeConversation({ c: i, m: notify.created.m, tags })
 									);
 									// eslint-disable-next-line @typescript-eslint/ban-ts-comment
 									// @ts-ignore
-									dispatch(handleAddMessagesInConversation(msgsReference));
+									dispatch(handleNotifyCreatedConversations(conversations));
+								}
+								if (notify.created.m) {
+									dispatch(handleCreatedMessages({ m: notify.created.m }));
+									dispatch(handleCreatedMessagesInConversation({ m: notify.created.m }));
 								}
 							}
+							if (notify.modified) {
+								const soapModifiedConversations = notify.modified.c;
+								if (soapModifiedConversations) {
+									const conversations = map(soapModifiedConversations, (i) =>
+										normalizeConversation({ c: i, tags })
+									);
+									// eslint-disable-next-line @typescript-eslint/ban-ts-comment
+									// @ts-ignore
+									dispatch(handleNotifyModifiedConversations(keyBy(conversations, 'id')));
+									updateConversationsOnly(normalizeConversations(soapModifiedConversations, tags));
+								}
+								const soapModifiedMessages = notify.modified.m;
+								if (soapModifiedMessages) {
+									const messages = map(soapModifiedMessages, (obj) =>
+										normalizeMailMessageFromSoap(obj)
+									);
+									// eslint-disable-next-line @typescript-eslint/ban-ts-comment
+									// @ts-ignore
+									// dispatch(handleModifiedMessages(messages));
+									updateMessagesOnly(messages);
+
+									// the condition filters messages with parent property (the only ones we need to update)
+									const toUpdate = filter(messages, 'parent');
+									if (toUpdate?.length > 0) {
+										// this function updates messages' parent in conversations. If parent never changes it does not need to be called
+										// eslint-disable-next-line @typescript-eslint/ban-ts-comment
+										// @ts-ignore
+										dispatch(handleModifiedMessagesInConversation(toUpdate));
+									}
+									// the condition filters messages with conversation property (the only ones we need to add to conversation)
+									const conversationToUpdate = filter(messages, 'conversation');
+									if (conversationToUpdate?.length > 0) {
+										const msgsReference = reduce(
+											conversationToUpdate,
+											(
+												acc: Array<{
+													id: string;
+													parent: string | undefined;
+													date: number | undefined;
+													conversation: Conversation;
+												}>,
+												msg: any
+											) => {
+												if (messagesState?.[msg?.id]) {
+													return [
+														...acc,
+														{
+															id: messagesState?.[msg?.id].id,
+															parent: messagesState?.[msg?.id].parent,
+															date: messagesState?.[msg?.id].date,
+															conversation: msg.conversation
+														}
+													];
+												}
+												return acc;
+											},
+											[]
+										);
+										// eslint-disable-next-line @typescript-eslint/ban-ts-comment
+										// @ts-ignore
+										dispatch(handleAddMessagesInConversation(msgsReference));
+									}
+								}
+							}
+							if (notify.deleted) {
+								deleteConversations(notify.deleted);
+								deleteMessages(notify.deleted);
+								dispatch(handleNotifyDeletedConversations(notify.deleted));
+								dispatch(handleDeletedMessages(notify.deleted));
+								dispatch(handleDeletedMessagesInConversation(notify.deleted));
+							}
+							setSeq(notify.seq);
 						}
-						if (notify.deleted) {
-							deleteConversations(notify.deleted);
-							deleteMessages(notify.deleted);
-							dispatch(handleNotifyDeletedConversations(notify.deleted));
-							dispatch(handleDeletedMessages(notify.deleted));
-							dispatch(handleDeletedMessagesInConversation(notify.deleted));
-						}
-						setSeq(notify.seq);
 					}
 				});
 			}